Move MessageSink class to own file. Move FileSink into message_sinks header
