+++ include/vcpkg/base/fwd/message_sinks.h
@@ -1,0 +1,13 @@
+#pragma once
+
+namespace vcpkg
+{
+    struct MessageSink;
+
+    extern MessageSink& null_sink;
+    extern MessageSink& stdout_sink;
+    extern MessageSink& stderr_sink;
+
+    struct FileSink;
+    struct CombiningSink;
+}
