+++ include/vcpkg/base/messages.h
@@ -371,84 +371,6 @@ namespace vcpkg::msg
 
 namespace vcpkg
 {
-    struct MessageSink
-    {
-        virtual void print(Color c, StringView sv) = 0;
-
-        void println() { this->print(Color::none, "\n"); }
-        void print(const LocalizedString& s) { this->print(Color::none, s); }
-        void println(Color c, const LocalizedString& s)
-        {
-            this->print(c, s);
-            this->print(Color::none, "\n");
-        }
-        inline void println(const LocalizedString& s)
-        {
-            this->print(Color::none, s);
-            this->print(Color::none, "\n");
-        }
-
-        template<class Message, class... Ts>
-        typename Message::is_message_type print(Message m, Ts... args)
-        {
-            this->print(Color::none, msg::format(m, args...));
-        }
-
-        template<class Message, class... Ts>
-        typename Message::is_message_type println(Message m, Ts... args)
-        {
-            this->print(Color::none, msg::format(m, args...).append_raw('\n'));
-        }
-
-        template<class Message, class... Ts>
-        typename Message::is_message_type print(Color c, Message m, Ts... args)
-        {
-            this->print(c, msg::format(m, args...));
-        }
-
-        template<class Message, class... Ts>
-        typename Message::is_message_type println(Color c, Message m, Ts... args)
-        {
-            this->print(c, msg::format(m, args...).append_raw('\n'));
-        }
-
-        void println_warning(const LocalizedString& s);
-        template<class Message, class... Ts>
-        typename Message::is_message_type println_warning(Message m, Ts... args)
-        {
-            println_warning(msg::format(m, args...));
-        }
-
-        void println_error(const LocalizedString& s);
-        template<class Message, class... Ts>
-        typename Message::is_message_type println_error(Message m, Ts... args)
-        {
-            println_error(msg::format(m, args...));
-        }
-
-        template<class Message, class... Ts, class = typename Message::is_message_type>
-        LocalizedString format_warning(Message m, Ts... args)
-        {
-            return format(msg::msgWarningMessage).append(m, args...);
-        }
-
-        template<class Message, class... Ts, class = typename Message::is_message_type>
-        LocalizedString format_error(Message m, Ts... args)
-        {
-            return format(msg::msgErrorMessage).append(m, args...);
-        }
-
-        MessageSink(const MessageSink&) = delete;
-        MessageSink& operator=(const MessageSink&) = delete;
-
-    protected:
-        MessageSink() = default;
-        ~MessageSink() = default;
-    };
-
-    extern MessageSink& null_sink;
-    extern MessageSink& stdout_sink;
-    extern MessageSink& stderr_sink;
 
     DECLARE_MESSAGE(ABaseline, (), "", "a baseline");
     DECLARE_MESSAGE(ABaselineObject, (), "", "a baseline object");
