+++ src/vcpkg/base/files.cpp
@@ -2034,38 +2034,26 @@ namespace vcpkg
         return ExpectedL<ReadFilePointer>{std::move(ret)};
     }
 
+    WriteFilePointer Filesystem::open_for_write(const Path& file_path, Append append, LineInfo li)
+    {
+        std::error_code ec;
+        auto ret = this->open_for_write(file_path, append, ec);
+        if (ec)
+        {
+            exit_filesystem_call_error(li, ec, __func__, {file_path});
+        }
+
+        return ret;
+    }
+
     WriteFilePointer Filesystem::open_for_write(const Path& file_path, std::error_code& ec)
     {
-        return open_for_write(file_path, WriteFilePointer::Append::NO, ec);
+        return open_for_write(file_path, Append::NO, ec);
     }
 
     WriteFilePointer Filesystem::open_for_write(const Path& file_path, LineInfo li)
     {
-        std::error_code ec;
-        auto ret = this->open_for_write(file_path, ec);
-        if (ec)
-        {
-            exit_filesystem_call_error(li, ec, __func__, {file_path});
-        }
-
-        return ret;
-    }
-
-    WriteFilePointer Filesystem::open_for_appending(const Path& file_path, std::error_code& ec)
-    {
-        return open_for_write(file_path, WriteFilePointer::Append::YES, ec);
-    }
-
-    WriteFilePointer Filesystem::open_for_appending(const Path& file_path, LineInfo li)
-    {
-        std::error_code ec;
-        auto ret = this->open_for_appending(file_path, ec);
-        if (ec)
-        {
-            exit_filesystem_call_error(li, ec, __func__, {file_path});
-        }
-
-        return ret;
+        return open_for_write(file_path, Append::NO, li);
     }
 
     struct RealFilesystem final : Filesystem
@@ -2632,7 +2620,7 @@ namespace vcpkg
                                  const std::vector<std::string>& lines,
                                  std::error_code& ec) override
         {
-            vcpkg::WriteFilePointer output{file_path, WriteFilePointer::Append::NO, ec};
+            vcpkg::WriteFilePointer output{file_path, Append::NO, ec};
             if (!ec)
             {
                 for (const auto& line : lines)
@@ -3318,7 +3306,7 @@ namespace vcpkg
         virtual void write_contents(const Path& file_path, StringView data, std::error_code& ec) override
         {
             StatsTimer t(g_us_filesystem_stats);
-            auto f = open_for_write(file_path, WriteFilePointer::Append::NO, ec);
+            auto f = open_for_write(file_path, Append::NO, ec);
             if (!ec)
             {
                 auto count = f.write(data.data(), 1, data.size());
@@ -3584,9 +3572,7 @@ namespace vcpkg
             return ReadFilePointer{file_path, ec};
         }
 
-        virtual WriteFilePointer open_for_write(const Path& file_path,
-                                                WriteFilePointer::Append append,
-                                                std::error_code& ec) override
+        virtual WriteFilePointer open_for_write(const Path& file_path, Append append, std::error_code& ec) override
         {
             return WriteFilePointer{file_path, append, ec};
         }
