+++ src/vcpkg/build.cpp
@@ -3,6 +3,7 @@
 #include <vcpkg/base/chrono.h>
 #include <vcpkg/base/hash.h>
 #include <vcpkg/base/json.h>
+#include <vcpkg/base/message_sinks.h>
 #include <vcpkg/base/messages.h>
 #include <vcpkg/base/optional.h>
 #include <vcpkg/base/stringview.h>
@@ -973,28 +974,13 @@ namespace vcpkg
         }
 
         const BuildInfo build_info = read_build_info(fs, paths.build_info_file_path(action.spec));
-        const size_t error_count = [&]() {
-            struct FileSink : MessageSink
-            {
-                WriteFilePointer out_file;
-                Path stdoutlog;
-                FileSink(Filesystem& fs, StringView stdoutlog)
-                    : m_stdoutlog(stdoutlog), m_out_file(fs.open_for_appending(m_stdoutlog, VCPKG_LINE_INFO))
-                {
-                }
-                void print(Color c, StringView sv) override
-                {
-                    msg::write_unlocalized_text_to_stdout(c, sv);
-                    Checks::msg_check_exit(VCPKG_LINE_INFO,
-                                           out_file.write(sv.data(), 1, sv.size()) == sv.size(),
-                                           msgErrorWhileWriting,
-                                           msg::path = stdoutlog);
-                }
-            };
-            FileSink file_sink{fs, stdoutlog};
-            return perform_post_build_lint_checks(
-                action.spec, paths, pre_build_info, build_info, scfl.source_location, file_sink);
-        }();
+        size_t error_count = 0;
+        {
+            FileSink file_sink{fs, stdoutlog, Append::YES};
+            CombiningSink combo_sink{stdout_sink, file_sink};
+            error_count = perform_post_build_lint_checks(
+                action.spec, paths, pre_build_info, build_info, scfl.source_location, combo_sink);
+        };
 
         auto find_itr = action.feature_dependencies.find("core");
         Checks::check_exit(VCPKG_LINE_INFO, find_itr != action.feature_dependencies.end());
