+++ src/vcpkg/commands.bootstrap-standalone.cpp
@@ -1,3 +1,5 @@
+#include <vcpkg/base/fwd/message_sinks.h>
+
 #include <vcpkg/base/downloads.h>
 #include <vcpkg/base/optional.h>
