+++ src/vcpkg/export.chocolatey.cpp
@@ -1,3 +1,5 @@
+#include <vcpkg/base/fwd/message_sinks.h>
+
 #include <vcpkg/base/system.process.h>
 
 #include <vcpkg/commands.h>
